+++ plom/produce/buildNamedPDF.py
@@ -8,6 +8,7 @@ import os
 from pathlib import Path
 from multiprocessing import Pool
 from tqdm import tqdm
+import csv
 
 from .mergeAndCodePages import make_PDF, make_fakePDF
 from . import paperdir
@@ -30,7 +31,50 @@ def _make_PDF(x):
         make_PDF(*y)
 
 
-def build_all_papers(spec, global_page_version_map, classlist, fakepdf=False):
+def outputProductionCSV(spec, make_PDF_args):
+    """Output a csv with info on produced papers. Take the make_PDF_args that were used and dump them in a csv
+
+    Arguments:
+        spec (dict): exam specification, see :func:`plom.SpecVerifier`.
+        make_PDF_args (list): a list of tuples of info for each paper
+    """
+    # a tuple in make_pdf_args is a tuple
+    # 0 - spec["name"],
+    # 1 - spec["publicCode"],
+    # 2 - spec["numberOfPages"],
+    # 3 - spec["numberOfVersions"],
+    # 4 - paper_index,
+    # 5 - page_version = dict(page:version)
+    # 6 - student_info = dict(id:sid ,name:sname)
+    # we only need the last 3 of these
+    numberOfPages = spec["numberOfPages"]
+    numberOfQuestions = spec["numberOfQuestions"]
+
+    header = ["test_number", "sID", "sname"]
+    for q in range(1, numberOfQuestions + 1):
+        header.append("q{}.version".format(q))
+    for p in range(1, numberOfPages + 1):
+        header.append("p{}.version".format(p))
+    with open("produced_papers.csv", "w") as csvfile:
+        csv_writer = csv.writer(csvfile)
+        csv_writer.writerow(header)
+        for paper in make_PDF_args:
+            if paper[6]:  # if named paper then give id and name
+                row = [paper[4], paper[6]["id"], paper[6]["name"]]
+            else:  # just skip those columns
+                row = [paper[4], None, None]
+            for q in range(1, numberOfQuestions + 1):
+                # get first page of question to infer version
+                p = spec["question"]["{}".format(q)]["pages"][0]
+                row.append(paper[5][p])
+            for p in range(1, numberOfPages + 1):
+                row.append(paper[5][p])
+            csv_writer.writerow(row)
+
+
+def build_all_papers(
+    spec, global_page_version_map, classlist, fakepdf=False, no_qr=False
+):
     """Builds the papers using _make_PDF.
 
     Based on `numberToName` this uses `_make_PDF` to create some
@@ -78,7 +122,8 @@ def build_all_papers(spec, global_page_version_map, classlist, fakepdf=False):
                 paper_index,
                 page_version,
                 student_info,
-                fakepdf,
+                no_qr,
+                fakepdf,  # should be last
             )
         )
 
@@ -88,6 +133,9 @@ def build_all_papers(spec, global_page_version_map, classlist, fakepdf=False):
     num_PDFs = len(make_PDF_args)
     with Pool() as pool:
         r = list(tqdm(pool.imap_unordered(_make_PDF, make_PDF_args), total=num_PDFs))
+    # output CSV with all this info in it
+    print("Writing produced_papers.csv.")
+    outputProductionCSV(spec, make_PDF_args)
 
 
 def confirm_processed(spec, msgr, classlist):