+++ CHANGELOG.md
@@ -11,8 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+* `plom-build make` now outputs a csv file showing the test numbers, question/page versions and student-info if named paper.
+* `plom-build make --noqr` builds pdf files without QR codes stamped on them.
 
 ### Fixed
+* Adjust-Pages: fix deduping when shared page not included in current question.
+* Fixed some times in manager tool.
 
 
 ## [0.5.10] - 2020-12-06